fix((UI-991): strip atlassian name from connections integration to work in triggers forms
