+++ src/components/organisms/triggers/formParts/fileAndFunction.tsx
@@ -8,7 +8,7 @@ import { eventTypesPerIntegration } from "@src/constants/triggers";
 import { TriggerTypes } from "@src/enums";
 import { SelectOption } from "@src/interfaces/components";
 import { useCacheStore } from "@src/store";
-import { stripGoogleConnectionName } from "@src/utilities";
+import { stripAtlassianConnectionName, stripGoogleConnectionName } from "@src/utilities";
 import { TriggerFormData } from "@validations";
 
 import { ErrorMessage, Input } from "@components/atoms";
@@ -54,10 +54,12 @@ export const TriggerSpecificFields = ({
 			return;
 		}
 
-		const connectionIntegration = stripGoogleConnectionName(
-			connections
-				?.find((connection) => connection.connectionId === connectionId)
-				?.integrationName?.toLowerCase() ?? ""
+		const connectionIntegration = stripAtlassianConnectionName(
+			stripGoogleConnectionName(
+				connections
+					?.find((connection) => connection.connectionId === connectionId)
+					?.integrationName?.toLowerCase() ?? ""
+			)
 		);
 
 		if (
