+++ src/enums/components/index.ts
@@ -6,6 +6,7 @@ export {
 	defaultGoogleConnectionName,
 	fitleredIntegrationsMap,
 	HiddenIntegrationsForTemplates,
+	defaultAtlassianConnectionName,
 } from "@enums/components/connection.enum";
 export { DrawerName } from "@enums/components/drawer.enum";
 export { InputVariant } from "@enums/components/input.enum";
